Added new audio classes
