--- source/AudioEffects.cpp
@@ -1,156 +1,0 @@
-#include "AudioEffects.h"
-
-CAudio::CAudio(const std::string& soundFile)
-{
-	// assuming sound file *Not music*	
-	// check if file exists
-	if(FileExists(soundFile))
-	{
-		// File sucessfully loaded
-		std::cout << "Audio file loaded...\nFILE DIR: " << soundFile << "\n\n";
-
-		// Manage SFML resources
-		buffer.loadFromFile(soundFile);
-		sound.setBuffer(buffer);
-
-		sound.setVolume(40.0f); // 0 to 100
-		sound.setPitch(4.0f);
-		sound.setLoop(false);
-		sound.setPosition(soundPos);
-
-		// setting up sound position
-		soundPos.x = 0.0;
-		soundPos.y = 0.0;
-		soundPos.z = 0.0;
-
-		// setting up sound velocity
-		soundVelocity.x = 0.0;
-		soundVelocity.y = 0.0;
-		soundVelocity.z = 0.0;
-
-		// setting up lister position
-		listenerPos.x = 0.0;
-		listenerPos.y = 0.0;
-		listenerPos.z = 0.0;
-
-		// setting up listener forward
-		listenerForward.x =  0.0;
-		listenerForward.y =  0.0;
-		listenerForward.z = -1.0;
-
-		// setting up listener up
-		listenerUp.x = 0.0;
-		listenerUp.y = 1.0;
-		listenerUp.z = 0.0;
-
-		// 0 to 100
-		sf::Listener::setGlobalVolume(100.0f);
-		sf::Listener::setPosition(listenerPos);
-		sf::Listener::setDirection(listenerForward);
-		sf::Listener::setUpVector(listenerUp);
-
-		// audio won't loop
-		sound.setLoop(false);
-	}
-	else
-	{
-		std::cerr << "ERROR: Cannot find audio file \nFILE DIR: " << soundFile <<  "\n\n";
-	}
-}
-
-void CAudio::Play()
-{
-	sound.play();
-}
-
-void CAudio::Stop()
-{
-
-}
-
-void CAudio::PauseUnPauseFile()
-{
-	if(paused)
-	{
-		paused = false;
-		sound.play();
-	}
-	else
-	{
-		paused = true;
-		sound.pause();
-	}
-}
-
-
-bool CAudio::FileExists(const std::string& soundFile)
-{
-	std::ifstream infile(soundFile);
-	return	 infile.good();
-}
-
-float Timer()
-{
-	static clock_t currTime = clock();
-	clock_t prevTime = currTime;
-	currTime = clock();
-	return float(currTime - prevTime) / CLOCKS_PER_SEC;
-}
-
-//#ifdef  _MINI_GAME
-//
-//sf::SoundBuffer bufferMini;
-//sf::Sound soundMini;
-//float pitch = 1.0f;
-//
-//#endif
-
-//#ifdef _MINI_GAME
-//void LoadMiniSound()
-//{
-//	if (!bufferMini.loadFromFile(".\\AudioFiles\\synthOneShot.wav"))
-//	{
-//		std::cout << "Error loading sound file" << std::endl;
-//		while (!_kbhit());
-//		return;
-//	}
-//
-//	soundMini.setBuffer(bufferMini);
-//
-//	// Set the properties of the source. Details of all available properties are in the SFML documentation of the Sound class
-//	soundMini.setVolume(40.0f); // 0 to 100
-//	soundMini.setPitch(pitch);
-//	soundMini.setLoop(false);
-//	soundMini.setPosition(0.0, 0.0, 0.0);
-//}
-//
-//void PlayMiniSound()
-//{
-//	soundMini.play();
-//}
-//
-//void IncrementPitch()
-//{
-//	if (pitch < 4.0f)
-//	{
-//		pitch = ChangeSemitone(pitch, 1.0f);
-//		soundMini.setPitch(pitch);
-//	}	
-//}
-//
-//void DecrementPitch()
-//{
-//	if (pitch > 0.117f)
-//	{
-//		pitch = ChangeSemitone(pitch, -1.0f);
-//		soundMini.setPitch(pitch);
-//	}	
-//}
-//
-//float ChangeSemitone(float frequency, float variation)
-//{
-//	static float semitone_ratio = pow(2.0f, 1.0f / 12.0f);
-//	return frequency * pow(semitone_ratio, variation);
-//}
-//
-//#endif
